+++ vkcs/internal/services/iam/s3accounts/results.go
@@ -2,6 +2,8 @@ package s3accounts
 
 import (
 	"github.com/gophercloud/gophercloud"
+	"github.com/gophercloud/gophercloud/pagination"
+	paginationutil "github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/util/pagination"
 )
 
 // CreateResult is the result of a create request. Call its Extract method
@@ -17,6 +19,18 @@ func (r CreateResult) Extract() (*CreateS3AccountResponse, error) {
 	return &s, err
 }
 
+// ExtractS3Accounts extracts and returnsa S3Account slice.
+func ExtractS3Accounts(r pagination.Page) ([]S3Account, error) {
+	var s []S3Account
+	err := ExtractS3AccountsInto(r, &s)
+	return s, err
+}
+
+// ExtractS3AccountsInto converts a page into a S3Account slice.
+func ExtractS3AccountsInto(r pagination.Page, v any) error {
+	return r.(S3AccountPage).Result.ExtractIntoSlicePtr(v, "data")
+}
+
 // GetResult is the result of a get request. Call its Extract method
 // to interpret a result as a S3Account.
 type GetResult struct {
@@ -28,6 +42,15 @@ func (r GetResult) Extract() (*S3Account, error) {
 	var s S3Account
 	err := r.ExtractInto(&s)
 	return &s, err
+}
+
+func NewS3AccountPage(r pagination.PageResult) S3AccountPage {
+	return S3AccountPage{OffsetPageBase: paginationutil.OffsetPageBase{PageResult: r, Label: "data"}}
+}
+
+// S3AccountPage is Pager that is returned from a call to the List function.
+type S3AccountPage struct {
+	paginationutil.OffsetPageBase
 }
 
 type DeleteResult struct {