+++ CHANGELOG.md
@@ -8,6 +8,7 @@ description: |-
 # VKCS Provider's changelog
 #### v0.14.0 (unreleased)
 - Add vkcs_iam_s3_account resource
+- Add vkcs_iam_s3_account data source
 
 #### v0.13.2
 - Add details to the vkcs_compute_instance creation error